native-host/requirements.txt
pyobjc-framework-Cocoa>=10.3.1
pyobjc-framework-Quartz>=10.3.1
pandas>=2.1.4
reportlab>=4.0.0
